requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

loguru==0.7.3
numpy==2.2.5
openai==1.76.0
pydantic==2.11.3
python-dotenv==1.1.0
serpapi==0.1.5
srsly==2.5.1
streamlit==1.44.1
tqdm==4.67.1
